+++ drawio-custom-plugins/src/index.ts
@@ -1,3 +1,4 @@
 import "./linkSelectedNodeWithData";
 import "./liveshare";
 import "./focus";
+import "./menu-entries";