+++ src/Extension.ts
@@ -7,11 +7,8 @@ import { DrawioWebviewInitializer } from "./DrawioWebviewInitializer";
 import { DrawioEditorManager } from "./DrawioEditorManager";
 import { LinkCodeWithSelectedNodeService } from "./features/CodeLinkFeature";
 import { EditDiagramAsTextFeature } from "./features/EditDiagramAsTextFeature";
-import { autorun } from "mobx";
 import { LiveshareFeature } from "./features/LiveshareFeature";
 import { InsiderFeedbackFeature } from "./features/InsiderFeedbackFeature";
-
-const drawioChangeThemeCommand = "hediet.vscode-drawio.changeTheme";
 
 export class Extension {
 	public readonly dispose = Disposable.fn();
@@ -36,10 +33,6 @@ export class Extension {
 	private readonly drawioWebviewInitializer = new DrawioWebviewInitializer(
 		this.config,
 		this.log
-	);
-
-	private readonly statusBar = this.dispose.track(
-		vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Right)
 	);
 
 	constructor(private readonly packageJsonPath: string) {
@@ -67,133 +60,5 @@ export class Extension {
 				}
 			)
 		);
-
-		this.dispose.track(
-			vscode.commands.registerCommand(drawioChangeThemeCommand, () =>
-				this.changeTheme()
-			)
-		);
-
-		this.dispose.track(
-			vscode.commands.registerCommand(
-				"hediet.vscode-drawio.convert",
-				() => this.convert()
-			)
-		);
-
-		this.dispose.track(
-			vscode.commands.registerCommand("hediet.vscode-drawio.export", () =>
-				this.export()
-			)
-		);
-
-		this.dispose.track({
-			dispose: autorun(
-				() => {
-					const activeEditor = this.editorManager.activeDrawioEditor;
-					this.statusBar.command = drawioChangeThemeCommand;
-
-					if (activeEditor) {
-						this.statusBar.text = `Theme: ${activeEditor.config.theme}`;
-						this.statusBar.show();
-					} else {
-						this.statusBar.hide();
-					}
-				},
-				{ name: "Update UI" }
-			),
-		});
-	}
-
-	private async convert(): Promise<void> {
-		const activeDrawioEditor = this.editorManager.activeDrawioEditor;
-		if (!activeDrawioEditor) {
-			return;
-		}
-
-		const result = await vscode.window.showQuickPick(
-			[
-				{
-					label: ".drawio.svg",
-					description: "Converts the diagram to an editable SVG file",
-				},
-				{
-					label: ".drawio",
-					description: "Converts the diagram to a drawio file",
-				},
-
-				{
-					label: ".drawio.png",
-					description: "Converts the diagram to an editable png file",
-				},
-			].filter((x) => x.label !== activeDrawioEditor.fileExtension)
-		);
-
-		if (!result) {
-			return;
-		}
-		await activeDrawioEditor.convertTo(result.label);
-	}
-
-	private async export(): Promise<void> {
-		const activeDrawioEditor = this.editorManager.activeDrawioEditor;
-		if (!activeDrawioEditor) {
-			return;
-		}
-
-		const result = await vscode.window.showQuickPick([
-			{
-				label: ".svg",
-				description: "Exports the diagram to a SVG file",
-			},
-			{
-				label: ".png",
-				description: "Exports the diagram to a png file",
-			},
-			{
-				label: ".drawio",
-				description: "Exports the diagram to a drawio file",
-			},
-		]);
-
-		if (!result) {
-			return;
-		}
-		await activeDrawioEditor.exportTo(result.label);
-	}
-
-	private async changeTheme(): Promise<void> {
-		const activeDrawioEditor = this.editorManager.activeDrawioEditor;
-		if (!activeDrawioEditor) {
-			return;
-		}
-
-		let availableThemes = ["automatic", "min", "atlas", "dark", "Kennedy"];
-
-		const originalTheme = activeDrawioEditor.config.theme;
-		availableThemes = availableThemes.filter((t) => t !== originalTheme);
-		availableThemes.unshift(originalTheme);
-
-		const result = await vscode.window.showQuickPick(
-			availableThemes.map((theme) => ({
-				label: theme,
-				description: `Selects Theme "${theme}"`,
-				theme,
-			})),
-			{
-				onDidSelectItem: async (item) => {
-					await activeDrawioEditor.config.setTheme(
-						(item as any).theme
-					);
-				},
-			}
-		);
-
-		if (!result) {
-			await activeDrawioEditor.config.setTheme(originalTheme);
-			return;
-		}
-
-		await activeDrawioEditor.config.setTheme(result.theme);
 	}
 }